+++ src/Entities/Composer.php
@@ -2,15 +2,32 @@
 
 namespace Maestriam\Samurai\Entities;
 
-use Maestriam\Samurai\Contracts\ComposerContract;
-
-class Composer implements ComposerContract
+class Composer extends Source
 {
+    /**
+     * Descrição do tema
+     */
     private string $desc;
 
-    public function __construct(Vendor $vendor, string $desc = null)
+    /**
+     * Nome do template
+     */
+    protected string $template = 'composer';
+    
+    /**
+     * Caminho absoluto do arquivo
+     */
+    private string $path;
+
+    /**
+     * Instância com as regras de negócio sobre composer.json
+     *
+     * @param Vendor $vendor
+     * @param string $desc
+     */
+    public function __construct(Theme $theme, string $desc = null)
     {
-        $this->setVendor($vendor);
+        $this->setTheme($theme);
 
         if ($desc) {
             $this->description($desc);
@@ -34,16 +51,48 @@ class Composer implements ComposerContract
     }
 
     /**
-     * Define o vendor do tema
+     * Executa a criação de um composer.json dentro do tema
      *
-     * @param Vendor $vendor
      * @return Composer
      */
-    private function setVendor(Vendor $vendor) : Composer
+    public function create() : Composer
     {
-        $this->vendor = $vendor;
+        $info = $this->createFile();
+
+        $this->setPath($info->absolute_path);
 
         return $this;
+    }
+
+    /**
+     * Retorna o caminho do arquivo composer.json dentro do tema
+     *
+     * @return string|null
+     */
+    public function path() : ?string
+    {
+        return $this->path;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    protected function placeholders() : array
+    {
+        return [
+            'description' => $this->description(),
+            'authorName'  => $this->theme->author()->name(),
+            'authorEmail' => $this->theme->author()->email(),
+            'package'     => $this->theme->vendor()->package(),
+        ];
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    protected function filename() : string
+    {
+        return 'composer.json';
     }
 
     /**
@@ -60,12 +109,30 @@ class Composer implements ComposerContract
     }
 
     /**
-     * Retorna a descrição do tema
+     * Retorna a descrição do tema  
+     * Se não houver uma descrição definida, retorna a descrição padrão
+     * inserida no arquivo config.php 
      *
      * @return string
      */
     private function getDescription() : string
     {
-        return $this->desc;
+        return $this->desc ?? $this->config()->description();
+    }
+
+    /**
+     * Define o caminho completo do arquivo composer.json dentro do tema
+     *
+     * @param string $path
+     * @return Composer
+     */
+    private function setPath(string $path) : Composer
+    {
+        if (! is_file($path)) {
+            throw new \Exception('');
+        }
+
+        $this->path = $path;
+        return $this;
     }
 }