+++ tests/Unit/Entities/Theme/MakeThemeTest.php
@@ -3,6 +3,7 @@
 namespace Maestriam\Samurai\Tests\Unit\Entities\Theme;
 
 use Maestriam\Samurai\Entities\Theme;
+use Maestriam\FileSystem\Support\FileSystem;
 
 /**
  * Testes de funcionalidades básicas apresentadas no README.md
@@ -11,10 +12,8 @@ class MakeThemeTest extends ThemeTestCase
 {
     public function testCreateTheme()
     {
-        $name = 'xuxu/pacoca';
+        $theme = new Theme('bands/ozzy');
 
-        $theme = new Theme($name);
-
-        $ret =  $theme->build();
+        $theme->make();
     }    
 }